Change: Move version CLI parser into an own module
The version helper module should just contain functions for handling
version changes.
