+++ pontos/version/__init__.py
@@ -18,7 +18,6 @@
 import sys
 from typing import NoReturn, Tuple, Type
 
-from pontos.version.helper import initialize_default_parser
 from pontos.version.version import VersionCommand
 
 from .__version__ import __version__
@@ -26,6 +25,7 @@ from .cmake import CMakeVersionCommand
 from .errors import VersionError
 from .go import GoVersionCommand
 from .javascript import JavaScriptVersionCommand
+from .parser import initialize_default_parser
 from .python import PythonVersionCommand
 
 COMMANDS: Tuple[Type[VersionCommand]] = (
