+++ pontos/version/helper.py
@@ -17,7 +17,6 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
 
-import argparse
 import re
 
 from packaging.version import InvalidVersion, Version
@@ -49,47 +48,6 @@ def is_version_pep440_compliant(version: str) -> bool:
     Checks if the provided version is a PEP 440 compliant version string
     """
     return version == safe_version(version)
-
-
-def initialize_default_parser() -> argparse.ArgumentParser:
-    """
-    Returns a default argument parser containing:
-    - verify
-    - show
-    - update
-    """
-    parser = argparse.ArgumentParser(
-        description="Version handling utilities.",
-        prog="version",
-    )
-    parser.add_argument(
-        "--quiet", help="don't print messages", action="store_true"
-    )
-
-    subparsers = parser.add_subparsers(
-        title="subcommands",
-        description="valid subcommands",
-        help="additional help",
-        dest="command",
-    )
-
-    verify_parser = subparsers.add_parser("verify")
-    verify_parser.add_argument("version", help="version string to compare")
-    subparsers.add_parser("show")
-
-    update_parser = subparsers.add_parser("update")
-    update_parser.add_argument("version", help="version string to use")
-    update_parser.add_argument(
-        "--force",
-        help="don't check if version is already set",
-        action="store_true",
-    )
-    update_parser.add_argument(
-        "--develop",
-        help="indicates if it is a develop version",
-        action="store_true",
-    )
-    return parser
 
 
 def safe_version(version: str) -> str:
